moved create ticket button
